+++ event-calendar/app/routes/Nav.js
@@ -1,7 +1,8 @@
 import React from "react";
+import { Link } from "@remix-run/react";
 
 const Nav = () => {
-  return <div>Nav</div>;
+  return <Link to="/calendar">Calendar</Link>;
 };
 
 export default Nav;
