+++ event-calendar/app/routes/calendar/index.tsx
@@ -1,8 +1,13 @@
+import { useLoaderData } from "@remix-run/react";
+import React from "react";
+import CalendarMain from "./Calendar.main";
+
 import type { LoaderFunctionArgs } from "@remix-run/node";
 import { json } from "@remix-run/node";
+
 import { db } from "~/db.server";
 
-export const eventsLoader = async ({ params }: LoaderFunctionArgs) => {
+export const loader = async ({ params }: LoaderFunctionArgs) => {
   return json(
     await db.event.findMany({
       where: {
@@ -17,3 +22,8 @@ export const eventsLoader = async ({ params }: LoaderFunctionArgs) => {
     })
   );
 };
+
+export default function Calendar() {
+  const events = useLoaderData<typeof loader>();
+  return <CalendarMain events={events} />;
+}
